caution and solde updating

+++ src/main/java/com/gmhis_backk/dto/AdmissionDTO.java
@@ -22,5 +22,6 @@ public class AdmissionDTO {
 	private Long service;
 	private Long act;
 	private Long practician;
-	private int caution;
+	private Double caution;
+
 }
